% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.null.test.R
\name{causalNullTest.control}
\alias{causalNullTest.control}
\title{Initialize control parameters for causalNullTest}
\usage{
causalNullTest.control(
  mu.SL.library = c("SL.mean", "SL.glm", "SL.gam", "SL.earth"),
  g.SL.library = c("SL.mean", "SL.glm", "SL.gam", "SL.earth"),
  g.n.bins = 2:(length(unique(A))/50),
  cross.fit = TRUE,
  V = 10,
  folds = NULL,
  save.nuis.fits = FALSE,
  mu.hat = NULL,
  g.hat = NULL,
  n.sim = 10000,
  return.Omega = FALSE,
  conf.level = 0.95,
  verbose = FALSE
)
}
\arguments{
\item{mu.SL.library}{Library of candidate learners for the outcome regression to be passed on to \code{\link[SuperLearner]{SuperLearner}}. Ignored if \code{mu.hats} is provided.}

\item{g.SL.library}{Library of candidate learners for the outcome regression to be passed on to \code{\link{cmdSuperLearner}}.}

\item{g.n.bins}{Numeric vector of number of bins to use for estimation of the propensity. Passed on to \code{\link{cmdSuperLearner}}.}

\item{cross.fit}{Logical indicating whether to cross-fit nuisance parameters. Defaults to \code{TRUE}.}

\item{V}{Positive integer number of folds for cross-fitting. Defaults to 10.}

\item{folds}{Optional \code{n x 1} numeric vector indicating which fold each observation is in.}

\item{save.nuis.fits}{Logical indicating whether to save the fitted nuisance objects.}

\item{mu.hat}{Optional pre-fit outcome regression. If \code{cross.fit} is \code{FALSE}, then a function that takes arguments \code{a} (a vector) and \code{w} (a data.frame) and returns predictions of the outcome regression function. If \code{cross.fit} is \code{TRUE}, then a list of functions of length \code{V} with the fitted outcome regression functions on each of the \code{V} training sets. If provided as a list of functions, then \code{folds} must be provided. If \code{mu.SL.library = NULL}, then \code{mu.hats} must be specified.}

\item{g.hat}{Optional pre-fit treatment propensities. If \code{cross.fit} is \code{FALSE}, then a function that takes arguments \code{a} (a vector) and \code{w} (a data.frame) and returns predictions of the standardized propensity function. If \code{cross.fit} is \code{TRUE}, then a list of functions of length \code{V} with the fitted standardized propensity functions on each of the \code{V} training sets. If provided as a list of functions, then \code{folds} must be provided. If \code{g.SL.library = NULL}, then \code{g.hats} must be specified.}

\item{n.sim}{Number of simulations to use for the limiting Gaussian process in computing approximate quantiles.}

\item{return.Omega}{Logical indicating whether to return the estimated primitive function Omega.}

\item{conf.level}{Optional confidence level to use for computing confidence bands for Omega.}

\item{verbose}{Logical indicating whether to print progress to the command line.}
}
\value{
Named list containing the control options.
}
\description{
This function initializes the control parameters for use in \code{causalNullTest}. The outcome regression function mu is by default estimated using \code{\link[SuperLearner]{SuperLearner}}, and the propensity is estimated using the conditional mixed density method implemented in \code{\link{cmdSuperLearner}}. Alternatively, the estimation process can be overriden by providing predictions from pre-fit nuisance estimators.
}
