% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal.null.test.R
\name{causalNullTest}
\alias{causalNullTest}
\title{Nonparametric test of flatness of a causal dose-response curve under exogeneity}
\usage{
causalNullTest(Y, A, W, p = 2, control = list())
}
\arguments{
\item{Y}{\code{n x 1} numeric vector of observed outcome values.}

\item{A}{\code{n x 1} numeric vector of exposure values.}

\item{W}{\code{n x p} data.frame of potential confounders.}

\item{p}{vector of exponents to use in the norm of Omega. Defaults to 2.}

\item{control}{Optional list of control parameters. See \code{\link{causalNullTest.control}} for details.}
}
\value{
\code{causalNullTest} returns a named list with the following elements:
\item{test}{A data.frame containing columns p (the exponent in the Lp norm used in the test), obs.stat (the observed Lp norm of the primitive Omega), p.val (the p-value corresponding to the test), ci.ll (the lower confidence limit corresponding to the Lp norm of the primitive), and ci.ul (the upper confidence limit corresponding to the Lp norm of the primitive).}
\item{control}{Controls used in fitting.}
If If \code{control$return.Omega == TRUE}, the following elements are also included in the output:
list(Omega.hat = data.frame(a=a.vals, Omega.hat), IF.vals = IF.vals, paths = paths))
\item{Omega.hat}{A data.frame with values of a and the corresponding estimated function Omega.hat(a).}
\item{IF.vals}{\code{n x m} matrix of influence function values of Omega.hat at every unique sorted value of the exposure. Each column is the influence function for a separate value of a, the values of which can be found in \code{Omega.hat}.}
\item{paths}{\code{control$n.sim x m} matrix of simulated paths from the limiting Gaussian process. Each row is an independent simulated path along the values of a.}
 If \code{save.nuis.fits = TRUE}, then the following elements are also included in the output:
 \item{mu.hat}{The estimated outcome regression function, as a list of fits if \code{control$cross.fit = TRUE}.}
 \item{g.hat}{The estimated propensity, possibly as a list of fits if \code{control$cross.fit = TRUE}.}
}
\description{
This function performs a hypothesis test that the causal dose-response curve theta(a) is flat on the support of the observed exposure A. The exposure may be discrete, continuous, or an arbitrary mixture of discrete and continuous components.  See the accompanying paper Westling (2020) for details.
}
\examples{
# Sample data
n <- 1000
W <- data.frame(W1 = runif(n))
Z <- rbinom(n, size = 1, prob = 1/(1 + exp(2-W$W1)))
A <- (1-Z) * rnorm(n, mean = W$W1, sd = abs(1 + W$W1))
Y <- rexp(n, rate = 1+abs(W$W1 * A))
causalNullTest(Y, A, W, p = c(1,2,Inf), control = list(cross.fit = FALSE, verbose=TRUE, g.n.bins = 2:5))
}
