% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixed.dens.R
\name{cmdSuperLearner.control}
\alias{cmdSuperLearner.control}
\title{Control parameters for the conditional mixed density Super Learner}
\usage{
cmdSuperLearner.control(
  n.bins = 2:floor(length(unique(A))/50),
  SL.library = c("SL.mean", "SL.glm", "SL.gam", "SL.earth"),
  saveFitLibrary = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{n.bins}{Vector of integers >= 2 indicating the number of bins to use for discretization. Defaults to \code{2:floor(n/100)}.}

\item{SL.library}{Library to use for bin-specific SuperLearners. Defaults to \code{c("SL.mean", "SL.glm", "SL.gam", "SL.earth")}.}

\item{saveFitLibrary}{Logical indicating whether to save the fit library on the full data. Defaults to \code{TRUE}. If \code{FALSE}, cannot obtain predicted values on new data later.}

\item{verbose}{Logical indicating whether to report progress of the estimation process.}
}
\value{
Returns a named list with control parameters.
}
\description{
This function initiates control parameters for the \code{\link{cmdSuperLearner}} function.
}
